refactor: move NotificationDataMapper to commonMain/alarm
