+++ shared/src/commonMain/kotlin/com/example/tasky/alarm/domain/mapper/NotificationDataMapper.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.alarm.domain.mapper
+package com.example.tasky.alarm.domain.mapper
 
 import com.example.tasky.agenda.domain.model.AgendaItem
 import com.example.tasky.alarm.domain.model.NotificationData
